examples/sdk/requirements.txt
tsbk
onnx
torch
mlflow-skinny